// src/utils/Types.ts
export type Weather = {
  accuracy: number | null
  altitude: null | number
  altitudeAccuracy: number | null
  heading: null | number
  latitude: number | null
  longitude: number | null
  speed: number | null
}

export type Country = {
  id: number | null
  name: String
  region: String
  country: String
}